+++ src/Console/Commands/CoreInstallCommand.php
@@ -31,11 +31,18 @@ class CoreInstallCommand extends Command
     {
         $this->alert($this->description);
 
-        $this->call('cortex:publish', ['--force' => $this->option('force'), '--resource' => $this->option('resource') ?: ['config']]);
+        $this->call('key:generate', ['--ansi' => true]);
+        $this->call('storage:link');
+
+        $this->call('self-diagnosis');
+
+        // Publish assets only if explicitly required, otherwise skip for clean installation
+        ! $this->option('resource') || $this->call('cortex:publish', ['--force' => $this->option('force'), '--resource' => $this->option('resource')]);
+
         $this->call('cortex:migrate', ['--force' => $this->option('force')]);
         $this->call('cortex:seed');
 
-        $this->call('cortex:autoload');
-        $this->call('cortex:activate');
+        $this->call('cortex:autoload', ['--force' => $this->option('force')]);
+        $this->call('cortex:activate', ['--force' => $this->option('force')]);
     }
 }